@@ -10,7 +10,7 @@
     "lint": ""
   },
   "engines": {
-    "node": "16"
+    "node": "18"
   },
   "main": "index.js",
   "dependencies": {
@@ -20,7 +20,7 @@
     "dotenv": "^16.0.1",
     "firebase-admin": "^9.12.0",
     "firebase-functions": "^3.22.0",
-    "firebase-tools": "^11.13.0",
+    "firebase-tools": "^11.29.0",
     "helmet": "^5.0.1",
     "jsonwebtoken": "^8.5.1",
     "jwks-rsa": "^2.0.5"
