@@ -3,8 +3,10 @@ const admin = require("firebase-admin");
 const axios = require("axios");
 const bucket = admin.storage().bucket();
 
-exports.scheduledFunction = functions.region("europe-west1").pubsub
-  .schedule("every 1 hours")
+exports.scheduledFunction = functions
+  .region("europe-west1")
+  .pubsub.schedule("every 1 hours")
+  .timeZone("Etc/GMT")
   .onRun(async (context) => {
     try {
       const server = await axios.get(
